fix: overall build fix #2

+++ tasks/mpi/ivanov_m_gauss_horizontal/src/ops_mpi.cpp
@@ -113,7 +113,7 @@ bool ivanov_m_gauss_horizontal_mpi::TestMPITaskParallel::validation() {
   if (world.rank() == 0) {
     auto* input_matrix = reinterpret_cast<double*>(taskData->inputs[0]);
     int size_of_matrix = taskData->inputs_count[0];
-    auto input_number_of_equations = *reinterpret_cast<size_t*>(taskData->inputs[1]);
+    auto input_number_of_equations = *reinterpret_cast<int*>(taskData->inputs[1]);
 
     if (input_matrix == nullptr) return false;
